composer-plugin: Don't output invalid versions in i18n-map (#22150)
"1.2.x-dev" is not valid semver. Replace anything that's invalid with
"0.0.0" so plugins with released versions will be preferred as the
source for i18n.

Also fix the attempted deletion of the i18n-map file if the plugin is
misconfigured.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1607989159